requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
python-dotenv==1.0.1
langchain==0.2.14
langchain-community==0.2.12
chromadb==0.5.3
openai==1.37.1
tiktoken==0.7.0
